requirements.txt
streamlit
sentence-transformers
scikit-learn
PyPDF2
python-docx
pandas
numpy
